More extensions (#5)
* fix: fix vscode extension permission

* chore: bump vscode extension version

* upgrade: upgrade @kksh/api and template ext build script.

Replace rollup with bun in worker template ext build.

* Use docker image with both bun and node in CI

* fix: improve build script to fix missing dist bug for template worker ext

* fix: add missing dep valibot to ext

* fix: change docker image used to build and upload extensions

* fix: docker run command in CI

* fix: Update dependencies in extensions to use @kksh/api v0.0.9

* Add base url for 3 custom ui extensions

https://github.com/kunkunsh/kunkun/commit/70cea0e003fca940ad474a7bea8abf2ad80d5a78

* ext: add disable-apple-quarantine extension

* feat: add system info extension, with battery info command

Need to wait for desktop v0.1.8 where plist parser api is added, and wait for @kksh/api 0.0.11 where system info API return types are fixed.

* feat: add icon for each battery info

* upgrade @kksh/api to 0.0.11, with fixed system info types

* feat: add periodic refresh for battery info command

* Add discord notification when new extension published

* Add DISCORD_WEBHOOK_URL env var for CI

* feat: add interval clear before goBack in battery info command

Need to wait for @kksh/api 0.0.12 publish. Currently using pnpm link.

* fix: wrong dev url in jwt ext and upgrade api

---------

+++ scripts/docker/entrypoint.sh
@@ -1,11 +1,15 @@
-cd /workspace
-rm *.tgz
+set -e
+# cd /workspace
+
+if [ -e *.tgz ]; then
+    rm *.tgz
+fi
 rm -rf node_modules
-cp -r /workspace /workspace-copy
-cd /workspace-copy
+export PATH=/root/.bun/bin:$PATH
 corepack enable pnpm
-pnpm i
+pnpm install
 pnpm run build
+npx @kksh/cli@latest verify --publish
 npm pack
 # check number of *.tgz file in current directory
 # if more than 1, then exit with error
@@ -14,3 +18,4 @@ if [ $(ls -1 *.tgz 2>/dev/null | wc -l) -gt 1 ]; then
     exit 1
 fi
 cp *.tgz /workspace
+rm -rf node_modules
